Update example to work with the latest FastAPI and add support for running in Docker
