--- Pipfile
@@ -1,14 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-flake8 = "*"
-
-[packages]
-fastapi = "*"
-uvicorn = "*"
-celery = "*"
-redis = "*"
-
