+++ README.md
@@ -1,29 +1,34 @@
-# fastapi with celery
+# FastAPI with Celery
 
-> Minimal example utilizing fastapi and celery with RabbitMQ for task queue, Redis for celery backend and flower for monitoring the celery tasks.
+> Minimal example utilizing FastAPI and Celery with RabbitMQ for task queue, Redis for Celery backend and flower for monitoring the Celery tasks.
 
 ## Requirements
 
-- Python >= 3.6.x
-  - pipenv (```pip install pipenv```)
 - Docker
-  - docker-compose
+  - [docker-compose](https://docs.docker.com/compose/install/)
 
-## Running
+## Run example
 
-### Install project dependencies
+1. Run command ```docker-compose up```to start up the RabbitMQ, Redis, flower and our application/worker instances.
+2. Navigate to the [http://localhost:8000/docs](http://localhost:8000/docs) and execute test API call. You can monitor the execution of the celery tasks in the console logs or navigate to the flower monitoring app at [http://localhost:5555](http://localhost:5555) (username: user, password: test).
 
-Execute the following command:
+## Run application/worker without Docker?
 
-```shell
-pipenv --python <your_python_version>
-pipenv install --dev
-```
+### Requirements/dependencies
 
-### Run example
+- Python >= 3.7
+  - [poetry](https://python-poetry.org/docs/#installation)
+- RabbitMQ instance
+- Redis instance
 
-1. Run command ```docker-compose up```to start up the RabbitMQ, Redis and flower instances.
-2. Start the fastapi web application with ```pipenv run uvicorn main:app --reload```.
-3. Start the celery worker with command ```pipenv run celery worker -A worker.celery_worker -l info -Q test-queue -c 1```
+> The RabbitMQ, Redis and flower services can be started with ```docker-compose -f docker-compose-services.yml up```
 
-Navigate to the [http://localhost:8000/docs](http://localhost:8000/docs) and execute test API call. You can monitor the execution of the celery tasks in the console logs or navigate to the flower monitoring app at [http://localhost:5555](http://localhost:5555) (username: user, password: test).
+### Install dependencies
+
+Execute the following command: ```poetry install --dev```
+
+### Run FastAPI app and Celery worker app
+
+1. Start the FastAPI web application with ```poetry run hypercorn app/main:app --reload```.
+2. Start the celery worker with command ```poetry run celery worker -A app.worker.celery_worker -l info -Q test-queue -c 1```
+3. Navigate to the [http://localhost:8000/docs](http://localhost:8000/docs) and execute test API call. You can monitor the execution of the celery tasks in the console logs or navigate to the flower monitoring app at [http://localhost:5555](http://localhost:5555) (username: user, password: test).
