+++ pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "fastapi-celery"
+version = "0.1.0"
+description = "FastAPI with Celery tasks minimal example"
+authors = ["Grega Vrbančič <grega.vrbancic@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.7"
+flake8 = "^3.8.3"
+fastapi = "^0.60.1"
+hypercorn = "^0.10.2"
+celery = "^4.4.6"
+redis = "^3.5.3"
+
+[tool.poetry.dev-dependencies]
+flake8 = "^3.8.3"
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
